requirements.txt
# Core dependencies
streamlit==1.29.0
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
pandas>=2.1.4
numpy>=1.25.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
Pillow>=10.4.0
prometheus-client>=0.19.0

# Visualization and analysis
matplotlib>=3.7.2
seaborn>=0.12.2
scipy>=1.11.3

# Machine learning
scikit-learn>=1.3.0
prophet>=1.1.4
statsmodels>=0.14.0
pmdarima>=2.0.3

# PDF generation
fpdf2>=2.5.6

# Development dependencies
pytest>=7.4.2
black>=23.10.1
flake8>=6.1.0

# Additional dependencies for compatibility
tqdm>=4.66.5
sqlalchemy>=2.0.36
pyarrow>=14.0.2
importlib-metadata>=6.11.0
packaging>=23.2